--- a/model/profile_test.go
+++ b/model/profile_test.go
@@ -87,7 +87,8 @@ func TestPprofProfileTransform(t *testing.T) {
 		},
 	}
 
-	output := pp.Transform(context.Background(), &transform.Config{DataStreams: true})
+	batch := &model.Batch{Profiles: []*model.PprofProfile{&pp}}
+	output := batch.Transform(context.Background(), &transform.Config{DataStreams: true})
 	require.Len(t, output, 2)
 	assert.Equal(t, output[0], output[1])