--- a/model/span_test.go
+++ b/model/span_test.go
@@ -175,10 +175,10 @@ func TestSpanTransform(t *testing.T) {
 	}
 
 	for _, test := range tests {
-		output := test.Span.Transform(context.Background(), &transform.Config{
+		output := test.Span.appendBeatEvents(context.Background(), &transform.Config{
 			DataStreams: true,
 			RUM:         transform.RUMConfig{SourcemapStore: &sourcemap.Store{}},
-		})
+		}, nil)
 		fields := output[0].Fields
 		assert.Equal(t, test.Output, fields, test.Msg)
 	}