## Motivation/summary

Rather than having each model type create a slice only to append it to another slice, append to that slice directly. Cuts out some unnecessary allocations.

Also, I have renamed and unexported the `Transform` methods on the event types to clarify that they are no longer transformed directly, but only ever through a `model.Batch`.

## How to test these changes

Non-functional change.

## Related issues

None.