@axw 2021-03-08
Modifying `processor/stream.BenchmarkBackendProcessor` to transform batches:

```
name                                          old time/op    new time/op    delta
BackendProcessor/heavy.ndjson/NoRateLimit-12    21.7ms ±21%    21.3ms ±12%    ~     (p=0.796 n=10+10)

name                                          old speed      new speed      delta
BackendProcessor/heavy.ndjson/NoRateLimit-12  18.6MB/s ±18%  18.9MB/s ±12%    ~     (p=0.781 n=10+10)

name                                          old alloc/op   new alloc/op   delta
BackendProcessor/heavy.ndjson/NoRateLimit-12    17.1MB ± 0%    17.0MB ± 0%  -0.70%  (p=0.000 n=10+10)

name                                          old allocs/op  new allocs/op  delta
BackendProcessor/heavy.ndjson/NoRateLimit-12      164k ± 0%      162k ± 0%  -1.16%  (p=0.000 n=10+10)
```